feat(server): add Stripe package and related resolvers  Added the Stripe package to the server-side package.json. Included the resolvers for Stripe that will get multiple donations (products) and one donation (product). Added the user resolver to retrieve the current user's information based on the user ID stored in the context on the server. Added the order function to fetch a specific order by its ID for an authenticated user. Added the checkout function that handles the creation of the Stripe checkout session. Added the addUser mutation which creates a new user with the provided arguments and generates a token for the user. Added the addOrder mutation which adds a new order to the authenticated user's list of orders. Added the updateUser mutation which updates the authenticated user's information with the provided arguments and returns the updated user object unless the user is not authenticated. Added the login mutation to handle user authentication, generating a token for the user and returning both the token and user object if successful.
feat(server): add Stripe package and related resolvers

Added the Stripe package to the server-side package.json. Included the resolvers for Stripe that will get multiple donations (products) and one donation (product). Added the user resolver to retrieve the current user's information based on the user ID stored in the context on the server. Added the order function to fetch a specific order by its ID for an authenticated user. Added the checkout function that handles the creation of the Stripe checkout session. Added the addUser mutation which creates a new user with the provided arguments and generates a token for the user. Added the addOrder mutation which adds a new order to the authenticated user's list of orders. Added the updateUser mutation which updates the authenticated user's information with the provided arguments and returns the updated user object unless the user is not authenticated. Added the login mutation to handle user authentication, generating a token for the user and returning both the token and user object if successful.