feat(models): add user.js model with bcrypt integration  Added the user.js model to include the bcrypt package to prehash the password of the user being created. The model includes firstName, lastName, email, and password documents. The order model is also included but is not complete yet and will be fixed later.
feat(models): add user.js model with bcrypt integration

Added the user.js model to include the bcrypt package to prehash the password of the user being created. The model includes firstName, lastName, email, and password documents. The order model is also included but is not complete yet and will be fixed later.