feat(client): add chakra-ui/react and update main.jsx  Added the package chakra-ui/react to the client side. Updated main.jsx to include the import for Chakra UI, and wrapped the RouterProvider with ChakraProvider as seen in the Chakra UI docs for easy access to the stylings offered through the package.
feat(client): add chakra-ui/react and update main.jsx

Added the package chakra-ui/react to the client side. Updated main.jsx to include the import for Chakra UI, and wrapped the RouterProvider with ChakraProvider as seen in the Chakra UI docs for easy access to the stylings offered through the package.